feat: enhance warp route extension tests
